fix(tests): add UserHandlerTest for CSV reading functionality and remove obsolete UserDatabaseHandlerTest and test data
