--- backend/users/src/test/resources/test-department-users.csv
@@ -1,4 +1,0 @@
-Department,Email
-Engineering,alice@example.org
-,bob@example.org
-Legal,carol@example.org
